requirements.txt
amqp 
asgiref 
async-timeout 
billiard 
black==24.4.0
celery 
chardet==5.2.0
click==8.1.8
click-didyoumean 
click-plugins 
click-repl 
Django 
djangorestframework 
isort==5.13.2
kombu 
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit 
psycopg2-binary 
python-dateutil 
python-dotenv==1.1.0
redis 
reportlab==4.0.8
six 
sqlparse 
tomli==2.2.1
typing_extensions==4.13.2
tzdata 
vine 
wcwidth 